fix(web): consolidate contest card
